# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onoot Boutique API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  discountPrice?: number | null;
  category: string;
  /** @nullable */
  subCategory?: string | null;
  /** @nullable */
  brand?: string | null;
  stock: number;
  images: string[];
  colors?: string[];
  tags?: string[];
  featured?: boolean;
  bestSeller?: boolean;
  rating: number;
  reviewCount: number;
  salesCount?: number;
  createdAt: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  discountPrice?: number;
  category: string;
  subCategory?: string;
  brand?: string;
  stock: number;
  images?: string[];
  colors?: string[];
  tags?: string[];
  featured?: boolean;
  bestSeller?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  /** @nullable */
  discountPrice?: number | null;
  category?: string;
  subCategory?: string;
  brand?: string;
  stock?: number;
  images?: string[];
  colors?: string[];
  tags?: string[];
  featured?: boolean;
  bestSeller?: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image?: string | null;
  productCount?: number;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  image?: string;
}

export interface CartItem {
  productId: number;
  quantity: number;
  price: number;
  product: Product;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  totalAmount: number;
  totalItems: number;
}

export interface CartItemInput {
  sessionId: string;
  productId: number;
  quantity: number;
  color?: string;
}

export interface CartItemUpdate {
  sessionId: string;
  quantity: number;
}

export interface CartItemRemove {
  sessionId: string;
}

export interface OrderItem {
  productId: number;
  quantity: number;
  price: number;
  productName: string;
  /** @nullable */
  productImage?: string | null;
}

export interface ShippingAddress {
  fullName: string;
  phone: string;
  address: string;
  city: string;
  country: string;
  /** @nullable */
  postalCode?: string | null;
}

export interface Order {
  id: number;
  userId: number;
  items: OrderItem[];
  totalAmount: number;
  paymentMethod: string;
  orderStatus: string;
  shippingAddress: ShippingAddress;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface OrderInput {
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  sessionId?: string | null;
  items: OrderItem[];
  totalAmount?: number;
  paymentMethod: string;
  shippingAddress: ShippingAddress;
  notes?: string;
}

export interface OrderStatusUpdate {
  orderStatus: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  createdAt: string;
  orderCount?: number;
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserUpdate {
  name?: string;
  phone?: string;
  avatar?: string;
  address?: string;
  city?: string;
  country?: string;
}

export interface Review {
  id: number;
  productId: number;
  /** @nullable */
  userId?: number | null;
  userName: string;
  rating: number;
  comment: string;
  createdAt: string;
}

export interface ReviewInput {
  userId?: number;
  userName: string;
  rating: number;
  comment: string;
}

export type AdminStatsOrdersByStatusItem = {
  status: string;
  count: number;
};

export type AdminStatsRevenueByMonthItem = {
  month: string;
  revenue: number;
};

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  totalUsers: number;
  recentOrders: Order[];
  ordersByStatus: AdminStatsOrdersByStatusItem[];
  revenueByMonth: AdminStatsRevenueByMonthItem[];
  newUsersThisMonth?: number;
}

export interface ProductSalesStat {
  product: Product;
  salesCount: number;
  revenue: number;
}

export type ListProductsParams = {
category?: string;
brand?: string;
search?: string;
minPrice?: number;
maxPrice?: number;
minRating?: number;
inStock?: boolean;
featured?: boolean;
bestSeller?: boolean;
onSale?: boolean;
page?: number;
limit?: number;
sortBy?: string;
};

export type GetCartParams = {
sessionId: string;
};

export type ListOrdersParams = {
userId?: number;
status?: string;
page?: number;
limit?: number;
};

export type GetTopProductsParams = {
limit?: number;
};

export type ListAdminUsersParams = {
page?: number;
limit?: number;
};