# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Onoot Boutique API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user's cart
      parameters:
        - name: sessionId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemRemove"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ─── Orders ──────────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (filtered by user or all for admin)
      parameters:
        - name: userId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── Users / Auth ─────────────────────────────────────────────────────────────
  /auth/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, price, category, stock, images, rating, reviewCount, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        discountPrice: { type: ["number", "null"] }
        category: { type: string }
        subCategory: { type: ["string", "null"] }
        brand: { type: ["string", "null"] }
        stock: { type: integer }
        images: { type: array, items: { type: string } }
        colors: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string } }
        featured: { type: boolean }
        bestSeller: { type: boolean }
        rating: { type: number }
        reviewCount: { type: integer }
        salesCount: { type: integer }
        createdAt: { type: string }
...
    CartItem:
      type: object
      required: [productId, quantity, price, product]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        price: { type: number }
        product: { $ref: "#/components/schemas/Product" }
    Cart:
      type: object
      required: [sessionId, items, totalAmount, totalItems]
      properties:
        sessionId: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/CartItem" }
        totalAmount: { type: number }
        totalItems: { type: integer }
    CartItemInput:
      type: object
      required: [sessionId, productId, quantity]
      properties:
        sessionId: { type: string }
        productId: { type: integer }
        quantity: { type: integer }
        color: { type: string }
    CartItemUpdate:
      type: object
      required: [sessionId, quantity]
      properties:
        sessionId: { type: string }
        quantity: { type: integer }
    CartItemRemove:
      type: object
      required: [sessionId]
      properties:
        sessionId: { type: string }

    # ─── Order ──────────────────────────────────────────────────────────────────
    OrderItem:
      type: object
      required: [productId, quantity, price, productName]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        price: { type: number }
        productName: { type: string }
        productImage: { type: ["string", "null"] }
    Order:
      type: object
      required: [id, userId, items, totalAmount, paymentMethod, orderStatus, shippingAddress, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        totalAmount: { type: number }
        paymentMethod: { type: string }
        orderStatus: { type: string }
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    OrderListResponse:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items: { $ref: "#/components/schemas/Order" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    ShippingAddress:
      type: object
      required: [fullName, phone, address, city, country]
      properties:
        fullName: { type: string }
        phone: { type: string }
        address: { type: string }
        city: { type: string }
        country: { type: string }
        postalCode: { type: ["string", "null"] }
    OrderInput:
      type: object
      required: [items, paymentMethod, shippingAddress]
      properties:
        userId: { type: ["integer", "null"] }
        sessionId: { type: ["string", "null"] }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        totalAmount: { type: number }
        paymentMethod: { type: string }
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        notes: { type: string }
    OrderStatusUpdate:
      type: object
      required: [orderStatus]
      properties:
        orderStatus: { type: string }

    # ─── User / Auth ─────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string }
        phone: { type: ["string", "null"] }
        avatar: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        createdAt: { type: string }
        orderCount: { type: integer }
...
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string }
        password: { type: string }
        name: { type: string }
        phone: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user: { $ref: "#/components/schemas/User" }
        token: { type: string }
...