# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onoot Boutique API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { ShippingAddress } from './shippingAddress';

export interface Order {
  id: number;
  userId: number;
  items: OrderItem[];
  totalAmount: number;
  paymentMethod: string;
  orderStatus: string;
  shippingAddress: ShippingAddress;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}