# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onoot Boutique API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  discountPrice?: number | null;
  category: string;
  /** @nullable */
  subCategory?: string | null;
  /** @nullable */
  brand?: string | null;
  stock: number;
  images: string[];
  colors?: string[];
  tags?: string[];
  featured?: boolean;
  bestSeller?: boolean;
  rating: number;
  reviewCount: number;
  salesCount?: number;
  createdAt: string;
}