# lib/api-zod/src/generated/types/shippingAddress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onoot Boutique API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShippingAddress {
  fullName: string;
  phone: string;
  address: string;
  city: string;
  country: string;
  /** @nullable */
  postalCode?: string | null;
}